getting values - ref + formData
