+++ forms/src/App.jsx
@@ -1,15 +1,16 @@
 import Header from './components/Header.jsx';
 import Login from './components/Login.jsx';
+import Signup from './components/Signup.jsx';
 
 function App() {
-  return (
-    <>
-      <Header />
-      <main>
-        <Login />
-      </main>
-    </>
-  );
+	return (
+		<>
+			<Header />
+			<main>
+				<Signup />
+			</main>
+		</>
+	);
 }
 
 export default App;
